Fix sort and filter form init, refactor related code ...
... also fix some additional bugs and get rid of
'__latest__' pseudo-op key on the client-side
