+++ lib/controller/kontext.py
@@ -627,6 +627,9 @@ class Kontext(Controller):
                             'user_id', None) == self.session_get('user', 'id')
                 else:  # initial query => ownership is clear
                     tpl_data['user_owns_conc'] = True
+                if '__latest__' in tpl_data['conc_forms_args']:
+                    tpl_data['conc_forms_args'][op_id] = tpl_data['conc_forms_args']['__latest__']
+                    del tpl_data['conc_forms_args']['__latest__']
             else:
                 tpl_data['Q'] = []
                 tpl_data['conc_persistence_op_id'] = None
